pyproject.toml
[tool.poetry]
name = "music-link-converter"
version = "0.1.0"
description = "A tool to convert Tidal/Spotify links."
authors = ["bnjn <ben@bnjn.tech>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
requests = "^2.31.0"
python-dotenv = "^1.0.0"
urllib3 = "< 2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

